Split `test_datasets.py` into sub-modules
